refactor(asset): modularise asset/types
