+++ src/asset/types.ts
@@ -1,0 +1,245 @@
+// asset/types - Main asset types and functions
+// Part of Stardazed TX
+// (c) 2016 by Arthur Langereis - @zenmumbler
+// https://github.com/stardazed/stardazed-tx
+
+import { ArrayOfNumber } from "core/array";
+import { Float2, Float3, Float4, Float4x4 } from "math/primarray";
+import { MeshData } from "mesh/meshdata";
+import { VertexIndexMapping } from "mesh/builder";
+import { Texture, TextureDescriptor, UseMipMaps } from "render/texture";
+
+import { LightDescriptor } from "world/light";
+
+export interface Asset {
+	name: string;
+	userRef?: any;
+}
+
+
+export interface Texture2D extends Asset {
+	url?: URL;
+
+	useMipMaps?: UseMipMaps;
+	descriptor?: TextureDescriptor;
+	texture?: Texture;
+}
+
+
+export interface TextureCube extends Asset {
+	filePathPosX?: string;
+	filePathNegX?: string;
+	filePathPosY?: string;
+	filePathNegY?: string;
+	filePathPosZ?: string;
+	filePathNegZ?: string;
+
+	useMipMaps: UseMipMaps;
+	descriptor?: TextureDescriptor;
+	texture?: Texture;
+}
+
+
+export const enum MaterialFlags {
+	usesSpecular               = 0x00000001,
+	usesEmissive               = 0x00000002,
+	isTranslucent              = 0x00000004,
+
+	diffuseAlphaIsTransparency = 0x00000100,
+	diffuseAlphaIsOpacity      = 0x00000200,
+
+	normalAlphaIsHeight        = 0x00000800,
+
+	isSkinned                  = 0x00001000   // FIXME: hmm
+}
+
+
+export interface Material extends Asset {
+	flags: MaterialFlags;
+
+	baseColour: Float3;
+
+	specularColour: Float3;
+	specularIntensity: number;
+	specularExponent: number;
+
+	emissiveColour: Float3;
+	emissiveIntensity: number;
+
+	opacity: number; // 0: fully transparent, 1: fully opaque (default)
+	metallic: number; // 0: fully di-electric (default), 1: fully metallic
+	roughness: number; // 0: fully smooth (default), 1: fully rough
+	anisotropy: number; // 1..16
+
+	textureScale: Float2;
+	textureOffset: Float2;
+
+	albedoTexture?: Texture2D;	// TODO: change this to array of textures with typed channels
+	specularTexture?: Texture2D;
+	normalTexture?: Texture2D;
+	heightTexture?: Texture2D;
+	transparencyTexture?: Texture2D;
+	emissiveTexture?: Texture2D;
+
+	roughnessTexture?: Texture2D;
+	metallicTexture?: Texture2D;
+	ambientOcclusionTexture?: Texture2D;
+}
+
+
+export function makeMaterial(name?: string): Material {
+	return {
+		name: name || "",
+		flags: 0,
+
+		baseColour: [1, 1, 1],
+
+		specularColour: [0, 0, 0],
+		specularIntensity: 0,
+		specularExponent: 0,
+
+		emissiveColour: [0, 0, 0],
+		emissiveIntensity: 0,
+
+		textureScale: [1, 1],
+		textureOffset: [0, 0],
+
+		opacity: 1,
+		anisotropy: 1,
+		metallic: 0,
+		roughness: 0
+	};
+}
+
+
+export interface Mesh extends Asset {
+	readonly meshData: MeshData;
+	readonly indexMap?: VertexIndexMapping;
+}
+
+
+export interface Transform {
+	position: Float3;
+	rotation: Float4; // quat
+	scale: Float3;
+}
+
+export function makeTransform(): Transform {
+	return {
+		position: [0, 0, 0],
+		rotation: [0, 0, 0, 1],
+		scale: [1, 1, 1]
+	};
+}
+
+
+export interface WeightedVertexGroup extends Asset {
+	indexes: Int32Array | null;
+	weights: Float64Array | null;
+
+	bindPoseLocalTranslation: Float3 | null;
+	bindPoseLocalRotation: Float4 | null;
+	bindPoseLocalMatrix: Float4x4 | null;
+}
+
+export interface Skin extends Asset {
+	groups: WeightedVertexGroup[];
+}
+
+
+// -- DEPRECATED?
+export interface Joint {
+	root: boolean;
+}
+
+export const enum AnimationProperty {
+	None,
+
+	TranslationX,
+	TranslationY,
+	TranslationZ,
+	RotationX,
+	RotationY,
+	RotationZ,
+	ScaleX,
+	ScaleY,
+	ScaleZ
+}
+
+
+export interface AnimationKeyData {
+	times: ArrayOfNumber;
+	values: ArrayOfNumber;
+}
+
+
+export interface AnimationTrack {
+	animationName: string;
+	property: AnimationProperty;
+	key: AnimationKeyData;
+}
+// -- /DEPRECATED?
+
+
+export const enum TransformAnimationField {
+	None,
+	Translation,
+	Rotation,
+	Scale
+}
+
+export interface TransformAnimationTrack {
+	field: TransformAnimationField;
+	key: Float32Array;
+}
+
+export interface TransformAnimation {
+	tracks: TransformAnimationTrack[];
+}
+
+export interface JointAnimation extends TransformAnimation {
+	jointIndex: number;
+	jointName?: string;
+}
+
+export interface SkeletonAnimation extends Asset {
+	frameTime: number;
+	frameCount: number;
+	jointAnims: JointAnimation[];
+}
+
+
+// TODO: how do I handle models, lights, cameras and generic nodes?
+
+
+export interface Light {
+	// FIXME: to be changed
+	descriptor: LightDescriptor;
+}
+
+
+export interface Model extends Asset {
+	transform: Transform;
+	children: Model[];
+	parent?: Model;
+
+	// components
+	mesh?: Mesh;
+	materials?: Material[];
+	light?: Light;
+
+	// DEPRECATED?
+	joint?: Joint; // FBX and MD5
+	vertexGroup?: WeightedVertexGroup; // FBX
+	animations?: AnimationTrack[]; // FBX
+	// /DEPRECATED?
+}
+
+export function makeModel(name: string, ref?: any): Model {
+	return {
+		name: name,
+		userRef: ref,
+		transform: makeTransform(),
+		children: []
+	};
+}
